requirements.txt
gitpython
spur
numpy
Pyro4
scp
psutil
keras_applications
keras_preprocessing
